Refactored: Moved ActionController instrumentation from Rails into ActionPack integration.
